feat:smoothed polyfit notebook, modified optimizer
The modified optimizer class that I used for previous notebooks is pulled into
a separate file to be called now. The modified optimizer does not eliminate
all coefficients under the threshold. Instead, it removes coefficients one
by one until the desired number of terms is reached.

In the smoothed polyfit notebook, each block of noisy data is assumed to be
representable by a multivariate polynomial of order +1 relative to the order
of the derivatives expected, ie. it should be such that the highest order
derivative is continuous. This smoothed multivariate polyfit is then used
to calculate the different terms.